+++ asf_search/ASFSearchResults.py
@@ -1,6 +1,7 @@
 from collections import UserList
 from multiprocessing import Pool
 import json
+from typing import List
 from asf_search import ASFSession, ASFSearchOptions
 from asf_search.download.file_download_type import FileDownloadType
 from asf_search.exceptions import ASFSearchError
@@ -42,6 +43,21 @@ class ASFSearchResults(UserList):
     def jsonlite2(self):
         return results_to_jsonlite2(self)
 
+    def find_urls(self, extension: str = None, pattern: str = r'.*', directAccess: bool = False) -> List[str]:
+        """Returns a flat list of all https or s3 urls from all results matching an extension and/or regex pattern
+        param extension: the file extension to search for. (Defaults to `None`)
+            - Example: '.tiff'
+        param pattern: A regex pattern to search each url for.(Defaults to `False`)
+            - Example: `r'(QA_)+'` to find urls with 'QA_' at least once
+        param directAccess: should search in s3 bucket urls (Defaults to `False`)
+        """
+        urls = []
+
+        for product in self:
+            urls.extend(product.find_urls(extension=extension, pattern=pattern, directAccess=directAccess))
+        
+        return sorted(list(set(urls)))
+    
     def __str__(self):
         return json.dumps(self.geojson(), indent=2, sort_keys=True)