+++ asf_search/Products/S1Product.py
@@ -35,7 +35,7 @@ class S1Product(ASFStackableProduct):
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
 
-        self.properties['s3Urls'] = self._get_s3_urls()
+        self.properties['s3Urls'] = self._get_s3_uris()
 
         if self.has_baseline():
             self.baseline = self.get_baseline_calc_properties()
@@ -124,7 +124,7 @@ class S1Product(ASFStackableProduct):
         stack_opts.beamMode = [self.properties['beamModeType']]
         stack_opts.flightDirection = self.properties['flightDirection']
         stack_opts.relativeOrbit = [int(self.properties['pathNumber'])]  # path
-        stack_opts.platform = [PLATFORM.SENTINEL1A, PLATFORM.SENTINEL1B]
+        stack_opts.platform = [PLATFORM.SENTINEL1A, PLATFORM.SENTINEL1B, PLATFORM.SENTINEL1C]
 
         if self.properties['polarization'] in ['HH', 'HH+HV']:
             stack_opts.polarization = ['HH', 'HH+HV']