+++ asf_search/export/jsonlite.py
@@ -214,8 +214,9 @@ class JSONLiteStreamArray(list):
             if result[key] in ["NA", "NULL"]:
                 result[key] = None
 
-        if "temporalBaseline" in p.keys() or "perpendicularBaseline" in p.keys():
+        if "temporalBaseline" in p.keys():
             result["temporalBaseline"] = p["temporalBaseline"]
+        if "perpendicularBaseline" in p.keys():
             result["perpendicularBaseline"] = p["perpendicularBaseline"]
 
         if p.get("processingLevel") == "BURST":  # is a burst product
@@ -229,6 +230,16 @@ class JSONLiteStreamArray(list):
             if p.get('validityStartDate'):
                 result['opera']['validityStartDate'] = p.get('validityStartDate')
 
+        if p.get('platform') == 'NISAR':
+            result['nisar'] = {
+                'pgeVersion':  p.get('pgeVersion'),
+                'mainBandPolarization':  p.get('mainBandPolarization'),
+                'sideBandPolarization':  p.get('sideBandPolarization'),
+                'frameCoverage':  p.get('frameCoverage'),
+                'jointObservation':  p.get('jointObservation'),
+                'rangeBandwidth':  p.get('rangeBandwidth'),
+            }
+        
         return result
 
     def getOutputType(self) -> str: